Error	28	Constant value '-128' cannot be converted to a 'byte' (use 'unchecked' syntax to override)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Tools.cs	240	43	Novensys.eCard.SDK Error	35	Cannot convert type 'int' to 'bool'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\SesiuneCard.cs	289	114	Novensys.eCard.SDK Error	52	Syntax error, ':' expected	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\TerminalManager.cs	772	40	Novensys.eCard.SDK Error	53	Invalid expression term '^'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\TerminalManager.cs	776	7	Novensys.eCard.SDK Error	54	; expected	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\TerminalManager.cs	776	8	Novensys.eCard.SDK Error	55	Syntax error, ':' expected	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\TerminalManager.cs	779	52	Novensys.eCard.SDK Error	56	Invalid expression term '^'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\TerminalManager.cs	783	7	Novensys.eCard.SDK Error	57	; expected	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\TerminalManager.cs	783	8	Novensys.eCard.SDK Error	86	Control cannot fall through from one case label ('case '<' (0x3C):') to another	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\IO\XmlAsn1InputStream.cs	1459	13	Novensys.eCard.SDK Error	87	Control cannot fall through from one case label ('case '!' (0x21):') to another	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\IO\XmlAsn1InputStream.cs	1465	19	Novensys.eCard.SDK Error	97	Control cannot fall through from one case label ('case '!' (0x21):') to another	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\IO\XmlAsn1InputStream.cs	2447	21	Novensys.eCard.SDK Error	98	Control cannot fall through from one case label ('case '!' (0x21):') to another	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\IO\XmlAsn1InputStream.cs	2712	15	Novensys.eCard.SDK Error	99	Invalid expression term ')'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	878	7	Novensys.eCard.SDK Error	100	; expected	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	878	8	Novensys.eCard.SDK Error	101	Invalid expression term '{'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	879	7	Novensys.eCard.SDK Error	102	) expected	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	879	8	Novensys.eCard.SDK Error	103	{ expected	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	883	44	Novensys.eCard.SDK Error	106	Constant value '-128' cannot be converted to a 'byte' (use 'unchecked' syntax to override)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	644	39	Novensys.eCard.SDK Error	107	Constant value '-2147483648' cannot be converted to a 'uint' (use 'unchecked' syntax to override)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	714	19	Novensys.eCard.SDK Error	108	The name 'outRemainder' does not exist in the current context	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	856	9	Novensys.eCard.SDK Error	109	The name 'index2' does not exist in the current context	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	856	27	Novensys.eCard.SDK Error	110	The name 'num2' does not exist in the current context	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	856	54	Novensys.eCard.SDK Error	111	Only assignment, call, increment, decrement, await, and new object expressions can be used as a statement	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	878	7	Novensys.eCard.SDK Error	112	Cannot use local variable 'num4' before it is declared	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	880	9	Novensys.eCard.SDK Error	113	A local variable named 'num4' cannot be declared in this scope because it would give a different meaning to 'num4', which is already used in a 'child' scope to denote something else	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	881	15	Novensys.eCard.SDK Error	114	The name 'index2' does not exist in the current context	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\BigInteger.cs	883	27	Novensys.eCard.SDK Error	116	The type or namespace name 'GetType' does not exist in the namespace 'Novensys.ASN1.Type' (are you missing an assembly reference?)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Asn1Factory.cs	165	57	Novensys.eCard.SDK Error	118	The type or namespace name 'GetType' does not exist in the namespace 'Novensys.ASN1.Type' (are you missing an assembly reference?)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Asn1Factory.cs	262	14	Novensys.eCard.SDK Error	119	'Novensys.ASN1.Type' is a 'namespace' but is used like a 'type'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Asn1Factory.cs	262	76	Novensys.eCard.SDK Error	120	'Novensys.ASN1.Type' is a 'namespace' but is used like a 'type'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Asn1Factory.cs	277	7	Novensys.eCard.SDK Error	121	The type or namespace name 'GetType' does not exist in the namespace 'Novensys.ASN1.Type' (are you missing an assembly reference?)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Asn1Factory.cs	280	21	Novensys.eCard.SDK Error	122	'Novensys.ASN1.Type' is a 'namespace' but is used like a 'type'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Asn1Factory.cs	289	77	Novensys.eCard.SDK Error	123	'Novensys.ASN1.Type' is a 'namespace' but is used like a 'type'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Util\Asn1Factory.cs	302	69	Novensys.eCard.SDK Error	124	Constant value '-128' cannot be converted to a 'byte' (use 'unchecked' syntax to override)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\IO\Asn1BitOutputStream.cs	118	25	Novensys.eCard.SDK Error	133	Cannot call an abstract base member: 'Novensys.ASN1.Type.Asn1TypeReader.init(System.IO.Stream)'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Type\Asn1TypeBerReader.cs	498	7	Novensys.eCard.SDK Error	134	Cannot call an abstract base member: 'Novensys.ASN1.Type.Asn1TypeReader.init(System.IO.Stream)'	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Type\Asn1TypeBerReader.cs	668	7	Novensys.eCard.SDK Error	135	Constant value '-128' cannot be converted to a 'byte' (use 'unchecked' syntax to override)	C:\Users\george.bungarzescu\Desktop\abcd\CEAS-NOVENSYS\trunk\Novensys.ASN1\Type\Asn1TypeBerWriter.cs	705	38	Novensys.eCard.SDK

+++ Novensys.ASN1/Type/Asn1TypeBerWriter.cs
@@ -20,7 +20,11 @@ namespace Novensys.ASN1.Type
     private static byte[] EPILOGUE_OF_INDEFINITE_LENGTH = new byte[2];
     private static byte[] PROLOGUE_OF_INDEFINITE_LENGTH = new byte[1]
     {
-      (byte) sbyte.MinValue
+
+        //use unchecked
+        
+        (byte) uint.MinValue
+       
     };
     protected bool _isCanonicalizingTimeValues = true;
     protected bool _isIndefiniteLengthForm = false;
